backend/requirements.txt
fastapi
httpx
pydantic
pydantic-ai
logfire
uvicorn
python-dotenv 